feat: Added Users.UpdateMetadata method
This commit adds a new function on the Users,
called `UpdateMetadata`.
This is responsible for updating the metadata of the
given user. This operation is additive, which means
that you don't have to pass the whole metadata
object, just the part you want to update/add and it
will do the necessary action on its own.