feat: Allow organizations update to receive metadata
The UpdateMetadata function is only allow us to append or modify the params,the deletions can be done only by setting the value of a property to null. This adds extra complexity to the front end in order to perform deletion. We modify the Update organizations function in order to receive metadata so we will be able to delete easier.