feat: Modify Issuer verification for all possible scenarios
The js sdk [1] supports overriding the expected issuer by allowing the
sdk user to set a couple of options.

This makes the issuer verification possible for a couple of non-standard
ways, namely when the customer's fapi sits behind a proxy or when the
domain is not the primary one.

[1] https://github.com/clerkinc/javascript/blob/c7c6912f34874467bc74104690fe9f95491cc10d/packages/backend/src/tokens/interstitialRule.ts#L158-L170

Fix USR-388