Groundwork for v2 (#206)
* chore: Fresh start

Removed all files to prepare for a v2 rewrite.

* feat: Base for implementing API operations

Added core types, methods and interfaces for building API operations.
The package provides a basic Backend that can communicate with the Clerk
API. The Backend can handle APIRequest requests and APIResponse
responses.
Types are also provided for API resources and parameters, as well as
expected errors.

* feat: Add CI workflow

Added a new CI workflow for Github Actions that will vet, and lint the
code and run the test suite.
Linting is done with golangci-lint. Tests are run against all Go
versions from 1.19 onwards.