fix: Params interface (#210)
Changed the Queryable interface to Params. Instead of operating on a
url.Values object, implementing types should return a url.Values object.
When we set the params to the querystring, we'll use the new interface
method.
Added a custom implementation to join URL paths. The new implementation
will preserve any query parameters in the URL. We can now replace
url.JoinPath, which required us to use Go v1.19 and upwards.