feat: HTTP middleware
HTTP middleware for enhancing http.Handler with Clerk authentication.
The http package provides two middleware:
- WithHeaderAuthentication checks the Authorization header for a valid
  JWT and sets the active session claims in the request context.
- RequireHeaderAuthentication responds with 403 Forbidden and halts the
  handler execution chain if it's unable to detect valid session claims.
Added a jwt package which implements the core function to validate a
JWT. The function name is Verify and is used by the middleware to
validate the Bearer JWT.
Added a jwk package which provides operations for the JSON Web Keys API.
We need to be able to fetch the instance JWK in order to validate and
parse the Bearer token.
Added a dependency to go-jose/v3 package for performing operations on
JWTs.