feat: Generator for API functions
Added a quick and dirty generator that takes a <package>/client.go file
and turns each of the <package>.Client methods into functions.
This allows us to automatically generate package-level functions for
each API operation. We write the Client methods and then call go
generate to handle creating API operations on the package.
As an example, let's say we have a domain package and a domain.Client,
which supports a Create method. Usage would then be
  domain.NewClient().Create()
When we run go generate for all packages, we'll also get a
  domain.Create()
function which calls the Client method under the hood.