feat: Cache requests for JWKS on JWT verification
The jwt.Verify method needs to fetch the JSON Web Key Set from the
API in order to verify the session JWT's validity.
The jwt.Verify method is used in the http.WithHeaderAuthorization
middleware, which means that in an HTTP server context, the method will
executed for every request.
We're adding a caching layer for the JWKS when we verify the session
JWT. This way we can cache the JWKS response from the API for 1 hour.