feat: JWKS cache is the HTTP middleware's concern (#244)
The JSON Web Key Set for an instance is not likely to change often.
We're taking advantage of this fact and save requests to the Clerk API
when verifying a JWT, by caching the API's GET /v1/jwks responses.
In reality, caching is needed by the HTTP middleware. We don't want to
fetch the JWKS for every request that the middleware needs to handle.
This commit moves the caching layer to the http middleware. The cache is
repurposed to store JSON Web Keys under their respective KeyIDs.
The jwt.Verify function now requires the JSON Web Key to be present in
the parameters and it is the responsibility of the caller to retrieve
it.
This way we achieve a clearer separation of concerns. The jwt.Verify
function should not be responsible for fetching the JWK, nor cache it.
It's the http middleware that needs the caching layer.