feat: JWT parsing with custom claims
Replaced the CustomClaims parameter with a CustomClaimsConstructor
function when verifying a session JWT. The option is also available in
the HTTP middleware.
The constructor function will be called when the JWT is parsed,
producing a new struct instance instead of writing on a single instance.
The custom claims will be made available in the SessionClaims.Custom
field.