fix: Don't expose jose types
Made sure that no go-jose types are exported from our package. Had to
add more exported types for JWT claims, so that we can re-use go-jose's
unmarshaling features.
We now have the following types for JWT claims:
- RegisteredClaims These are 'standard' public JWT claims registered in
  IANA.
- Claims These are Clerk-specific private claims, like session ID and
  active organization ID.
- SessionClaims These are the claims that are included in Clerk issued
  JWTs. They include RegisteredClaims, Claims and any other Custom keys
  from custom JWT templates.
Another change introduced here is that the jwt.Decode function will now
expose the Key ID header, if the JWT contains it. We need access to the
kid header when implementing a custom JWT verification flow and
jwt.Decode already does most of the work to parse a token. It makes
sense for the function to return the Key ID as well.