feat: Clock interface as a time source (#263)
Verifying a session JWT includes time sensitive validations, like
checking token expiration, or that the token can be used now ('nbf'
claim).
The HTTP middleware caches JSON web keys for one hour. After that, the
cache is considered invalid and a new request to fetch the JSON web key
set will be issued.
These are both flows that depend on the current time and time
comparisons.
Various system architectures with a distributed nature might not be
synchronized. A clock skew is expected.
In such cases it's common to share a "time source" so that all nodes can
coordinate on what the perceived current time is.
This commit adds a new clerk.Clock interface which can be passed as a
dependency in the http middleware functions and the jwt.Verify function.
We also add a dependency to the clockwork project, in order to use a
fake clock for our tests.