feat: jwt.Verify fetches the JSON web key (#266)
In order to reduce friction when manually verifying a session JWT, the
jwt.Verify function will fetch the JSON web key if it's not passed in
the params.
Users can specify the jwks.Client that will be used to invoke the API
GET /v1/jwks method.
Exported the method which fetches the JSON web key, since it can be
re-used by the middleware. It can also be used for consumers who wish to
fetch the JSON web key once and cache it. Caching is recommended.