feat: add SessionCreatedEvent type with User object (#457)
## Overview

This pull request introduces a new event struct to the `webhookevents` package for handling session creation events. The new struct embeds the `clerk.Session` and includes a `clerk.User` field for easier access to user information in session-related webhooks.

End goal is to ensure congruency between our webhook event structures by using the Go SDK for source of truth of these structs when publishing the webhook. 

### New `webhookevents` package struct:

* Added `SessionCreatedEvent` struct to `webhookevents/session_created_event.go`, embedding `clerk.Session` and including a `clerk.User` field.

Signed-off-by: Kenton Duprey <kenton@clerk.dev>